requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==1.10.3
async-timeout==4.0.3
Authlib==1.3.1
azure-core==1.30.2
beautifulsoup4==4.8.2
blinker==1.8.2
blis==0.7.11
botocore==1.34.157
braintree==4.29.0
cachelib==0.13.0
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
compressed-rtf==1.0.6
confection==0.1.5
cryptography==43.0.0
cymem==2.0.8
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
extract-msg==0.28.7
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
fsspec==2024.6.1
git-filter-repo==2.45.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
IMAPClient==2.1.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msgspec==0.18.6
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==2.1.0
olefile==0.47
openai==1.37.1
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdfkit==1.0.0
pdfminer.six==20191110
pdfplumber==0.11.2
pillow==10.4.0
portalocker==2.10.1
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
#pywin32==306
PyYAML==6.0.1
redis==5.0.7
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.1.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.12.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.4.8
sympy==1.13.1
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.43.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.2.0
